requirements.txt
Flask-Migrate~=4.1.0
Werkzeug~=3.1.3
SQLAlchemy~=2.0.37
livereload~=2.7.1
Flask~=3.1.0
Flask-Login~=0.6.3
itsdangerous~=2.2.0
pillow~=11.1.0
WTForms~=3.2.1